+++ src/app/src/reducers/index.js
@@ -17,6 +17,7 @@ import UIReducer from './UIReducer';
 import FacilitiesReducer from './FacilitiesReducer';
 import FacilityCountReducer from './FacilityCountReducer';
 import FeatureFlagsReducer from './FeatureFlagsReducer';
+import ClientInfoReducer from './ClientInfoReducer';
 
 export default combineReducers({
     auth: AuthReducer,
@@ -30,4 +31,5 @@ export default combineReducers({
     facilities: FacilitiesReducer,
     facilityCount: FacilityCountReducer,
     featureFlags: FeatureFlagsReducer,
+    clientInfo: ClientInfoReducer,
 });